// frontend/src/shared/api/model/cityResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */
import type { CityNameResponse } from './cityNameResponse';

export interface CityResponse {
  id?: number;
  region?: string;
  image?: string;
  description?: string;
  latitude?: number;
  longitude?: number;
  createdAt?: string;
  createdBy?: number;
  names?: CityNameResponse[];
  placeCount?: number;
}