// frontend/src/shared/api/model/createCityRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */
import type { CityNameRequest } from './cityNameRequest';

export interface CreateCityRequest {
  /**
   * @minLength 0
   * @maxLength 100
   */
  region: string;
  /**
   * @minLength 0
   * @maxLength 500
   */
  image?: string;
  description?: string;
  latitude: number;
  longitude: number;
  /** @minItems 1 */
  names: CityNameRequest[];
}