// frontend/src/shared/api/model/cityNameRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Tarikma API
 * Backend API for Tarik.ma
 * OpenAPI spec version: v1
 */

export interface CityNameRequest {
  /**
   * @minLength 1
   * @pattern ^(en|fr|ar|amz)$
   */
  languageCode: string;
  /**
   * @minLength 0
   * @maxLength 200
   */
  name: string;
  primary?: boolean;
}